refactor(separated all GUI functions & listeners)
