+++ babylon/client/js/GUI.js
@@ -1,12 +1,14 @@
-import { createArtistSpotify } from './musicFunctions'
+import { createArtistSpotify, getComposer } from './musicFunctions'
+import checkForPort from './checkForPort'
 
 // CastorGUI
 let canvas = document.getElementById("renderCanvas");
 let css = "button {cursor:pointer;} #textDialog{margin:6px}";
 let options = { themeRoot: "./dist/", themeGUI: "default" };
 let guisystem = new CASTORGUI.GUIManager(canvas, css, options);
+let pickedCameraPosition, pickResult;
 
-export default function createGUI(composerData) {
+function createCastorGUI(composerData) {
   if (!composerData) return
 
   //store composer info
@@ -25,4 +27,39 @@ export default function createGUI(composerData) {
 
   //add spotify, takes in name & id to append player to
   createArtistSpotify(composerName, '#dialog_content')
+}
+
+//retrieve composer information & use to create the GUI
+export function createComposerGUI(evt) {
+  if (scene.GUI) {
+    document.body.removeChild(document.getElementById("dialog"))
+    scene.GUI = false;
+    return
+  }
+  pickResult = scene.pick(scene.pointerX, scene.pointerY)
+  const meshHit = pickResult.pickedMesh.name;
+  if (pickResult.distance > 3) {
+    return
+  }
+  if (checkForPort(meshHit) && !scene.GUI) {
+    getComposer(meshHit)
+      .then((res) => createCastorGUI(res.data))
+      .catch(console.error.bind(console))
+    scene.GUI = true;
+    pickedCameraPosition = Object.assign({}, scene.cameras[0].position)
+  }
+}
+export function removeComposerGUI(event) {
+
+  let keyCodes = event.keyCode === 87 || event.keyCode === 83 || event.keyCode === 65 || event.keyCode === 68 ||
+    event.keyCode === 37 || event.keyCode === 38 || event.keyCode === 39 || event.keyCode === 40;
+
+  if (pickedCameraPosition && keyCodes) {
+    let currentCameraPosition = scene.cameras[0].position
+    let distanceAway = BABYLON.Vector3.Distance(pickedCameraPosition, currentCameraPosition)
+    if (distanceAway > .5 && scene.GUI) {
+      document.body.removeChild(document.getElementById("dialog"))
+      scene.GUI = false
+    }
+  }
 }
